Tests/LibWeb/Text/expected/async-scrolling/pseudo-element-wheel-scroll.txt

wheel routing admission: accepted
wheel scroll admission: accepted
downward wheel target before wheel: non-viewport
scroll event count: 1
host scrollTop: 0
window scrollY: 0
upward wheel target after wheel: non-viewport
